Update firebase.py
firebase.py의  api key 및 여러 config를 추가하였습니다.
